+++ apis/keda/v1alpha1/scaledjob_types.go
@@ -21,6 +21,11 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
+const (
+	defaultScaledJobMaxReplicaCount = 100
+	defaultScaledJobMinReplicaCount = 0
+)
+
 // +genclient
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
@@ -114,19 +119,23 @@ func init() {
 // MaxReplicaCount returns MaxReplicaCount
 func (s ScaledJob) MaxReplicaCount() int64 {
 	if s.Spec.MaxReplicaCount != nil {
+		if s.Spec.MinReplicaCount != nil && *s.Spec.MinReplicaCount > *s.Spec.MaxReplicaCount {
+			return int64(*s.Spec.MaxReplicaCount)
+		}
 		return int64(*s.Spec.MaxReplicaCount) - s.MinReplicaCount()
 	}
 
-	return 100
+	return defaultScaledJobMaxReplicaCount
 }
 
 // MinReplicaCount returns MinReplicaCount
 func (s ScaledJob) MinReplicaCount() int64 {
 	if s.Spec.MinReplicaCount != nil {
-		if *s.Spec.MinReplicaCount > *s.Spec.MaxReplicaCount {
+		if s.Spec.MaxReplicaCount != nil &&
+			*s.Spec.MinReplicaCount > *s.Spec.MaxReplicaCount {
 			return int64(*s.Spec.MaxReplicaCount)
 		}
 		return int64(*s.Spec.MinReplicaCount)
 	}
-	return 0
+	return defaultScaledJobMinReplicaCount
 }